refactor: 홈 라우팅 주소 /로 변경
