+++ src/app/search/page.tsx
@@ -1,0 +1,8 @@
+
+export default function HomeSearch() {
+  return (
+    <div className="flex min-h-screen items-center justify-center bg-zinc-50 font-sans dark:bg-black">
+      home/search
+    </div>
+  );
+}
